use 16bit for data that include last byte information

+++ source/driver-models/I2C.cpp
@@ -68,7 +68,7 @@ int I2C::stop()
 * @param data The byte to write.
 * @return DEVICE_OK on success, DEVICE_I2C_ERROR if the the write request failed.
 */
-int I2C::write(uint8_t data)
+int I2C::write(uint16_t data)
 {
     return DEVICE_NOT_IMPLEMENTED;
 }
